Consolidate sync-office-staff into sync-office-staff.ts
Jira ticket: CAMS-
