+++ backend/function-apps/dataflows/import/sync-office-staff.test.ts
@@ -1,11 +1,11 @@
 import { Timer } from '@azure/functions';
-import timerTrigger from './office-staff-sync.function';
-import { createMockAzureFunctionContext } from '../../../azure/testing-helpers';
-import { OfficesController } from '../../../../lib/controllers/offices/offices.controller';
-import { CamsError } from '../../../../lib/common-errors/cams-error';
-import { LoggerImpl } from '../../../../lib/adapters/services/logger.service';
+import { timerTrigger } from './sync-office-staff';
+import { createMockAzureFunctionContext } from '../../azure/testing-helpers';
+import { OfficesController } from '../../../lib/controllers/offices/offices.controller';
+import { CamsError } from '../../../lib/common-errors/cams-error';
+import { LoggerImpl } from '../../../lib/adapters/services/logger.service';
 
-describe('Office Staff Sync Function tests', () => {
+describe('Sync Office Staff Function tests', () => {
   const context = createMockAzureFunctionContext();
   const timer: Timer = {
     isPastDue: false,
