+++ backend/function-apps/dataflows/import/sync-office-staff.ts
@@ -1,0 +1,28 @@
+import { app, InvocationContext, Timer } from '@azure/functions';
+import ContextCreator from '../../azure/application-context-creator';
+import { OfficesController } from '../../../lib/controllers/offices/offices.controller';
+import { toAzureError } from '../../azure/functions';
+import { buildFunctionName } from '../dataflows-common';
+
+const MODULE_NAME = 'SYNC-OFFICE-STAFF';
+
+export async function timerTrigger(
+  _ignore: Timer,
+  invocationContext: InvocationContext,
+): Promise<void> {
+  const logger = ContextCreator.getLogger(invocationContext);
+  try {
+    const appContext = await ContextCreator.getApplicationContext({ invocationContext, logger });
+    const controller = new OfficesController();
+    await controller.handleTimer(appContext);
+  } catch (error) {
+    toAzureError(logger, MODULE_NAME, error);
+  }
+}
+
+export function setupSyncOfficeStaff() {
+  app.timer(buildFunctionName(MODULE_NAME, 'timerTrigger'), {
+    schedule: '0 0 5 * * *',
+    handler: timerTrigger,
+  });
+}
